Refactor gateway configuration and update documentation structure
- Updated proxy.py to use gateway_client_identity instead of separate client certificate and private key.
- Modified factories.py and test_computes.py to reflect changes in gateway configuration.
- Adjusted test_proxy.py to validate new client identity structure.
- Refactored react-router configuration to include new documentation paths and removed obsolete metadata files.
- Consolidated metadata definitions directly into respective documentation files for clarity and maintainability.
- Removed unused metadata files and updated imports accordingly in various components.
