+++ api/src/adapters/gateway.py
@@ -4,7 +4,6 @@ import tempfile
 from uuid import UUID
 from dataclasses import dataclass
 from collections.abc import AsyncIterator
-from src.models.gateways import USER_ID_HEADER, APPLICATION_ID_HEADER
 
 
 @dataclass(slots=True)
@@ -25,13 +24,12 @@ class GatewayResponse:
 class GatewayClient:
     """Send authenticated Platform requests to one compute gateway."""
 
-    def __init__(self, url: str, ca_certificate: str, client_certificate: str, client_private_key: str) -> None:
+    def __init__(self, url: str, ca_certificate: str, client_identity: str) -> None:
         """Initialize one gateway connection from persisted compute state."""
 
         self._url = url.rstrip("/")
         self._ca_certificate = ca_certificate
-        self._client_certificate = client_certificate
-        self._client_private_key = client_private_key
+        self._client_identity = client_identity
 
     async def request(
         self,
@@ -51,8 +49,8 @@ class GatewayClient:
         if query:
             url = f"{url}?{query}"
         headers = {
-            APPLICATION_ID_HEADER: str(application_id),
-            USER_ID_HEADER: str(user_id),
+            "x-longlink-application-id": str(application_id),
+            "x-user-id": str(user_id),
         }
         if content_type is not None:
             headers["content-type"] = content_type
@@ -60,7 +58,7 @@ class GatewayClient:
         # Authenticate the Platform using its client identity and trust only this Gateway CA.
         tls = ssl.create_default_context(cadata=self._ca_certificate)
         with tempfile.NamedTemporaryFile(mode="w") as identity:
-            identity.write(f"{self._client_certificate}\n{self._client_private_key}")
+            identity.write(self._client_identity)
             identity.flush()
             tls.load_cert_chain(identity.name)
         client = httpx2.AsyncClient(
