--- api/src/models/gateways.py
@@ -1,2 +1,0 @@
-APPLICATION_ID_HEADER = "x-longlink-application-id"
-USER_ID_HEADER = "x-user-id"
