--- web/src/platform/Home.metadata.ts
@@ -1,7 +1,0 @@
-export const metadata = {
-    path: '/',
-    title: 'LongLink',
-    seoTitle: 'LongLink | Python Platform for Business Applications',
-    description:
-        'LongLink is an open-source platform for building and running custom business-process applications with Python.',
-};
