+++ web/src/platform/Home.tsx
@@ -11,9 +11,14 @@ import { ClickableCard } from '@astryxdesign/core/ClickableCard';
 import { publicSeoMeta } from '@/lib/seo';
 import { Globe } from '@/components/Globe';
 import { PublicPage } from '@/layout/PublicPage';
-import { metadata } from './Home.metadata';
 
-export { metadata } from './Home.metadata';
+export const metadata = {
+    path: '/',
+    title: 'LongLink',
+    seoTitle: 'LongLink | Python Platform for Business Applications',
+    description:
+        'LongLink is an open-source platform for building and running custom business-process applications with Python.',
+};
 
 export const meta: MetaFunction = () => publicSeoMeta(metadata);
