--- web/src/platform/Pricing.metadata.ts
@@ -1,5 +1,0 @@
-export const metadata = {
-    path: '/pricing',
-    title: 'Pricing',
-    description: 'Simple LongLink pricing for building and running business-process applications.',
-};
