--- web/src/platform/docs/api/applications.metadata.ts
@@ -1,8 +1,0 @@
-export const metadata = {
-    path: '/docs/api/applications',
-    title: 'Applications',
-    description: 'Learn how LongLink registers, deploys, routes, and manages business applications.',
-    toc: [{ id: 'applications', label: 'Applications', level: 1 }],
-    lastUpdated: '2026-08-05',
-    editUrl: 'https://github.com/xLongLink/longlink/edit/main/web/src/platform/docs/api/applications.tsx',
-};
