+++ web/src/platform/docs/api/applications.tsx
@@ -7,7 +7,14 @@ import { Heading } from '@astryxdesign/core/Heading';
 
 const applicationPaths = ['Use', 'Adapt', 'Create'];
 
-export { metadata } from './applications.metadata';
+export const metadata = {
+    path: '/docs/api/applications',
+    title: 'Applications',
+    description: 'Learn how LongLink registers, deploys, routes, and manages business applications.',
+    toc: [{ id: 'applications', label: 'Applications', level: 1 }],
+    lastUpdated: '2026-08-05',
+    editUrl: 'https://github.com/xLongLink/longlink/edit/main/web/src/platform/docs/api/applications.tsx',
+};
 
 export default function ApplicationsDocumentation() {
     return (
