--- web/src/platform/docs/api/index.metadata.ts
@@ -1,9 +1,0 @@
-export const metadata = {
-    path: '/docs/api',
-    title: 'Overview',
-    seoTitle: 'Platform Documentation | LongLink',
-    description: 'Understand the LongLink Platform for organizations, applications, infrastructure, and operations.',
-    toc: [{ id: 'platform', label: 'Platform', level: 1 }],
-    lastUpdated: '2026-07-20',
-    editUrl: 'https://github.com/xLongLink/longlink/edit/main/web/src/platform/docs/api/index.tsx',
-};
