--- web/src/platform/docs/api/organizations.metadata.ts
@@ -1,11 +1,0 @@
-export const metadata = {
-    path: '/docs/api/organizations',
-    title: 'Organizations',
-    description: 'Learn how LongLink organizations, memberships, and access boundaries work.',
-    toc: [
-        { id: 'organizations', label: 'Organizations', level: 1 },
-        { id: 'users', label: 'Users', level: 2 },
-    ],
-    lastUpdated: '2026-07-25',
-    editUrl: 'https://github.com/xLongLink/longlink/edit/main/web/src/platform/docs/api/organizations.tsx',
-};
