+++ web/src/platform/docs/api/organizations.tsx
@@ -28,7 +28,17 @@ const organizationRoles = [
     },
 ];
 
-export { metadata } from './organizations.metadata';
+export const metadata = {
+    path: '/docs/api/organizations',
+    title: 'Organizations',
+    description: 'Learn how LongLink organizations, memberships, and access boundaries work.',
+    toc: [
+        { id: 'organizations', label: 'Organizations', level: 1 },
+        { id: 'users', label: 'Users', level: 2 },
+    ],
+    lastUpdated: '2026-07-25',
+    editUrl: 'https://github.com/xLongLink/longlink/edit/main/web/src/platform/docs/api/organizations.tsx',
+};
 
 export default function OrganizationsDocumentation() {
     return (
