--- web/src/platform/docs/index.metadata.ts
@@ -1,9 +1,0 @@
-export const metadata = {
-    path: '/docs',
-    title: 'Introduction',
-    seoTitle: 'Documentation | LongLink',
-    description: 'Learn how LongLink helps teams build and run structured business applications.',
-    toc: [{ id: 'introduction', label: 'Introduction', level: 1 }],
-    lastUpdated: '2026-07-10',
-    editUrl: 'https://github.com/xLongLink/longlink/edit/main/web/src/platform/docs/index.tsx',
-};
