+++ web/src/platform/docs/index.tsx
@@ -2,7 +2,15 @@ import { Text } from '@astryxdesign/core/Text';
 import { Stack } from '@astryxdesign/core/Stack';
 import { Heading } from '@astryxdesign/core/Heading';
 
-export { metadata } from './index.metadata';
+export const metadata = {
+    path: '/docs',
+    title: 'Introduction',
+    seoTitle: 'Documentation | LongLink',
+    description: 'Learn how LongLink helps teams build and run structured business applications.',
+    toc: [{ id: 'introduction', label: 'Introduction', level: 1 }],
+    lastUpdated: '2026-07-10',
+    editUrl: 'https://github.com/xLongLink/longlink/edit/main/web/src/platform/docs/index.tsx',
+};
 
 export default function IntroductionDocumentation() {
     return (
