--- web/src/platform/docs/sdk/building.metadata.ts
@@ -1,11 +1,0 @@
-export const metadata = {
-    path: '/docs/sdk/building',
-    title: 'Building',
-    description: 'Package LongLink applications into deployable images with metadata and environment requirements.',
-    toc: [
-        { id: 'building', label: 'Building', level: 1 },
-        { id: 'metadata', label: 'Metadata', level: 2 },
-    ],
-    lastUpdated: '2026-07-10',
-    editUrl: 'https://github.com/xLongLink/longlink/edit/main/web/src/platform/docs/sdk/building.tsx',
-};
