+++ web/src/platform/docs/sdk/building.tsx
@@ -5,7 +5,17 @@ import { Heading } from '@astryxdesign/core/Heading';
 import { CodeBlock } from '@astryxdesign/core/CodeBlock';
 import { Table, TableBody, TableCell, TableHeader, TableHeaderCell, TableRow } from '@astryxdesign/core/Table';
 
-export { metadata } from './building.metadata';
+export const metadata = {
+    path: '/docs/sdk/building',
+    title: 'Building',
+    description: 'Package LongLink applications into deployable images with metadata and environment requirements.',
+    toc: [
+        { id: 'building', label: 'Building', level: 1 },
+        { id: 'metadata', label: 'Metadata', level: 2 },
+    ],
+    lastUpdated: '2026-07-10',
+    editUrl: 'https://github.com/xLongLink/longlink/edit/main/web/src/platform/docs/sdk/building.tsx',
+};
 
 export default function BuildingDocumentation() {
     return (
