--- web/src/platform/docs/sdk/database.metadata.ts
@@ -1,14 +1,0 @@
-export const metadata = {
-    path: '/docs/sdk/database',
-    title: 'Database',
-    description: 'Use LongLink database helpers and migrations for application-owned data models.',
-    toc: [
-        { id: 'database', label: 'Database', level: 1 },
-        { id: 'basic-usage', label: 'Basic usage', level: 2 },
-        { id: 'timezone', label: 'Timezone', level: 2 },
-        { id: 'users-table', label: 'Users table', level: 2 },
-        { id: 'migrations', label: 'Migrations', level: 2 },
-    ],
-    lastUpdated: '2026-08-05',
-    editUrl: 'https://github.com/xLongLink/longlink/edit/main/web/src/platform/docs/sdk/database.tsx',
-};
