+++ web/src/platform/docs/sdk/database.tsx
@@ -38,7 +38,20 @@ const environments: EnvironmentRow[] = [
     },
 ];
 
-export { metadata } from './database.metadata';
+export const metadata = {
+    path: '/docs/sdk/database',
+    title: 'Database',
+    description: 'Use LongLink database helpers and migrations for application-owned data models.',
+    toc: [
+        { id: 'database', label: 'Database', level: 1 },
+        { id: 'basic-usage', label: 'Basic usage', level: 2 },
+        { id: 'timezone', label: 'Timezone', level: 2 },
+        { id: 'users-table', label: 'Users table', level: 2 },
+        { id: 'migrations', label: 'Migrations', level: 2 },
+    ],
+    lastUpdated: '2026-08-05',
+    editUrl: 'https://github.com/xLongLink/longlink/edit/main/web/src/platform/docs/sdk/database.tsx',
+};
 
 export default function DatabaseDocumentation() {
     return (
