--- web/src/platform/docs/sdk/environments.metadata.ts
@@ -1,11 +1,0 @@
-export const metadata = {
-    path: '/docs/sdk/environments',
-    title: 'Environments',
-    description: 'Configure LongLink applications for local development, testing, and production environments.',
-    toc: [
-        { id: 'environments', label: 'Environments', level: 1 },
-        { id: 'usage', label: 'Usage', level: 2 },
-    ],
-    lastUpdated: '2026-07-14',
-    editUrl: 'https://github.com/xLongLink/longlink/edit/main/web/src/platform/docs/sdk/environments.tsx',
-};
