+++ web/src/platform/docs/sdk/environments.tsx
@@ -4,7 +4,17 @@ import { Stack } from '@astryxdesign/core/Stack';
 import { Heading } from '@astryxdesign/core/Heading';
 import { CodeBlock } from '@astryxdesign/core/CodeBlock';
 
-export { metadata } from './environments.metadata';
+export const metadata = {
+    path: '/docs/sdk/environments',
+    title: 'Environments',
+    description: 'Configure LongLink applications for local development, testing, and production environments.',
+    toc: [
+        { id: 'environments', label: 'Environments', level: 1 },
+        { id: 'usage', label: 'Usage', level: 2 },
+    ],
+    lastUpdated: '2026-07-14',
+    editUrl: 'https://github.com/xLongLink/longlink/edit/main/web/src/platform/docs/sdk/environments.tsx',
+};
 
 export default function EnvironmentsDocumentation() {
     return (
