--- web/src/platform/docs/sdk/index.metadata.ts
@@ -1,14 +1,0 @@
-export const metadata = {
-    path: '/docs/sdk',
-    title: 'Overview',
-    seoTitle: 'Applications Documentation | LongLink',
-    description:
-        'Build LongLink applications locally with the Python SDK, XML pages, routes, storage, and database tools.',
-    toc: [
-        { id: 'application-sdk', label: 'Applications', level: 1 },
-        { id: 'create-a-project', label: 'Create a Project', level: 2 },
-        { id: 'local-development', label: 'Local Development', level: 2 },
-    ],
-    lastUpdated: '2026-07-10',
-    editUrl: 'https://github.com/xLongLink/longlink/edit/main/web/src/platform/docs/sdk/index.tsx',
-};
