+++ web/src/platform/docs/sdk/pages.tsx
@@ -75,7 +75,22 @@ const componentCategories: ComponentCategory[] = componentCategoryConfigurations
 
 const noop = () => undefined;
 
-export { metadata } from './pages.metadata';
+export const metadata = {
+    path: '/docs/sdk/pages',
+    title: 'Pages',
+    description: 'Build LongLink application pages with XML components, data bindings, and runtime metadata.',
+    toc: [
+        { id: 'pages', label: 'Pages', level: 1 },
+        { id: 'longlink-runtime-concepts', label: 'Runtime', level: 2 },
+        { id: 'longlink-state-elements', label: 'State', level: 2 },
+        { id: 'action', label: 'Action', level: 2 },
+        { id: 'content', label: 'Content', level: 2 },
+        { id: 'form', label: 'Form', level: 2 },
+        { id: 'layout', label: 'Layout', level: 2 },
+    ],
+    lastUpdated: '2026-07-21',
+    editUrl: 'https://github.com/xLongLink/longlink/edit/main/web/src/platform/docs/sdk/pages.tsx',
+};
 
 /** Renders one categorized group of XML page components. */
 function ComponentCategorySection({ category }: { category: ComponentCategory }) {
