+++ web/src/platform/docs/sdk/routes.tsx
@@ -5,7 +5,17 @@ import { Stack } from '@astryxdesign/core/Stack';
 import { Heading } from '@astryxdesign/core/Heading';
 import { CodeBlock } from '@astryxdesign/core/CodeBlock';
 
-export { metadata } from './routes.metadata';
+export const metadata = {
+    path: '/docs/sdk/routes',
+    title: 'Routes',
+    description: 'Add FastAPI routes to LongLink applications for APIs, actions, and process-specific behavior.',
+    toc: [
+        { id: 'routes', label: 'Routes', level: 1 },
+        { id: 'usage', label: 'Usage', level: 2 },
+    ],
+    lastUpdated: '2026-08-05',
+    editUrl: 'https://github.com/xLongLink/longlink/edit/main/web/src/platform/docs/sdk/routes.tsx',
+};
 
 export default function RoutesDocumentation() {
     return (
