--- web/src/platform/docs/sdk/storage.metadata.ts
@@ -1,12 +1,0 @@
-export const metadata = {
-    path: '/docs/sdk/storage',
-    title: 'Storage',
-    description: 'Use LongLink storage abstractions across local filesystems, tests, and production object storage.',
-    toc: [
-        { id: 'storage', label: 'Storage', level: 1 },
-        { id: 'usage', label: 'Usage', level: 2 },
-        { id: 'assets', label: 'Assets', level: 2 },
-    ],
-    lastUpdated: '2026-07-20',
-    editUrl: 'https://github.com/xLongLink/longlink/edit/main/web/src/platform/docs/sdk/storage.tsx',
-};
