--- web/src/platform/docs/sdk/testing.metadata.ts
@@ -1,11 +1,0 @@
-export const metadata = {
-    path: '/docs/sdk/testing',
-    title: 'Testing',
-    description: 'Test LongLink applications with isolated runtime configuration and Python testing workflows.',
-    toc: [
-        { id: 'testing', label: 'Testing', level: 1 },
-        { id: 'usage', label: 'Usage', level: 2 },
-    ],
-    lastUpdated: '2026-07-10',
-    editUrl: 'https://github.com/xLongLink/longlink/edit/main/web/src/platform/docs/sdk/testing.tsx',
-};
