--- web/src/platform/legal/impressum.metadata.ts
@@ -1,12 +1,0 @@
-export const metadata = {
-    path: '/impressum',
-    title: 'Impressum',
-    description: 'Read the LongLink legal notice and company information.',
-    toc: [
-        { id: 'impressum', label: 'Impressum', level: 1 },
-        { id: 'company', label: 'Company', level: 2 },
-        { id: 'contact', label: 'Contact', level: 2 },
-    ],
-    lastUpdated: '2026-07-06',
-    editUrl: 'https://github.com/xLongLink/longlink/edit/main/web/src/platform/legal/impressum.tsx',
-};
