+++ web/src/platform/legal/impressum.tsx
@@ -2,7 +2,19 @@ import { Link } from '@astryxdesign/core/Link';
 import { Text } from '@astryxdesign/core/Text';
 import { Stack } from '@astryxdesign/core/Stack';
 import { Heading } from '@astryxdesign/core/Heading';
-export { metadata } from './impressum.metadata';
+
+export const metadata = {
+    path: '/impressum',
+    title: 'Impressum',
+    description: 'Read the LongLink legal notice and company information.',
+    toc: [
+        { id: 'impressum', label: 'Impressum', level: 1 },
+        { id: 'company', label: 'Company', level: 2 },
+        { id: 'contact', label: 'Contact', level: 2 },
+    ],
+    lastUpdated: '2026-07-06',
+    editUrl: 'https://github.com/xLongLink/longlink/edit/main/web/src/platform/legal/impressum.tsx',
+};
 
 /** Renders the legal notice and company information. */
 export default function Impressum() {
