--- web/src/platform/legal/privacy.metadata.ts
@@ -1,24 +1,0 @@
-export const metadata = {
-    path: '/privacy',
-    title: 'Privacy',
-    description: 'Read the LongLink privacy policy.',
-    toc: [
-        { id: 'privacy-policy', label: 'Privacy Policy', level: 1 },
-        { id: 'scope-and-controller', label: '1. Scope and controller', level: 2 },
-        { id: 'our-roles', label: '2. Our roles', level: 2 },
-        { id: 'personal-data-we-process', label: '3. Personal data we process', level: 2 },
-        { id: 'why-we-process-personal-data', label: '4. Why we process personal data', level: 2 },
-        { id: 'service-providers-and-other-recipients', label: '5. Service providers and other recipients', level: 2 },
-        { id: 'international-transfers', label: '6. International transfers', level: 2 },
-        { id: 'cookies-and-similar-storage', label: '7. Cookies and similar storage', level: 2 },
-        { id: 'retention', label: '8. Retention', level: 2 },
-        { id: 'automated-operational-actions', label: '9. Automated operational actions', level: 2 },
-        { id: 'security-and-data-incidents', label: '10. Security and data incidents', level: 2 },
-        { id: 'your-rights', label: '11. Your rights', level: 2 },
-        { id: 'children', label: '12. Children', level: 2 },
-        { id: 'changes-to-this-policy', label: '13. Changes to this Policy', level: 2 },
-        { id: 'contact', label: '14. Contact', level: 2 },
-    ],
-    lastUpdated: '2026-07-20',
-    editUrl: 'https://github.com/xLongLink/longlink/edit/main/web/src/platform/legal/privacy.tsx',
-};
