--- web/src/platform/legal/terms.metadata.ts
@@ -1,32 +1,0 @@
-export const metadata = {
-    path: '/terms',
-    title: 'Terms of Service',
-    description: 'Read the LongLink terms of service.',
-    toc: [
-        { id: 'terms-of-service', label: 'Terms of Service', level: 1 },
-        { id: 'provider-acceptance-and-eligibility', label: '1. Provider, acceptance and eligibility', level: 2 },
-        { id: 'definitions-and-contract-documents', label: '2. Definitions and contract documents', level: 2 },
-        { id: 'service-and-beta-status', label: '3. Service and beta status', level: 2 },
-        { id: 'accounts-organizations-and-security', label: '4. Accounts, organizations and security', level: 2 },
-        { id: 'acceptable-use', label: '5. Acceptable use', level: 2 },
-        { id: 'plans-fees-and-managed-services', label: '6. Plans, fees and managed services', level: 2 },
-        {
-            id: 'refunds-and-consumer-cancellation-rights',
-            label: '7. Refunds and consumer cancellation rights',
-            level: 2,
-        },
-        { id: 'content-privacy-and-data-processing', label: '8. Content, privacy and data processing', level: 2 },
-        { id: 'our-technology-and-feedback', label: '9. Our technology and feedback', level: 2 },
-        { id: 'availability-and-third-party-services', label: '10. Availability and third-party services', level: 2 },
-        { id: 'warranties', label: '11. Warranties', level: 2 },
-        { id: 'limitation-of-liability', label: '12. Limitation of liability', level: 2 },
-        { id: 'indemnity-for-business-users', label: '13. Indemnity for business users', level: 2 },
-        { id: 'suspension-deletion-and-termination', label: '14. Suspension, deletion and termination', level: 2 },
-        { id: 'changes-to-these-terms', label: '15. Changes to these Terms', level: 2 },
-        { id: 'force-majeure', label: '16. Force majeure', level: 2 },
-        { id: 'general-and-governing-law', label: '17. General and governing law', level: 2 },
-        { id: 'contact', label: '18. Contact', level: 2 },
-    ],
-    lastUpdated: '2026-06-20',
-    editUrl: 'https://github.com/xLongLink/longlink/edit/main/web/src/platform/legal/terms.tsx',
-};
